Bug Fix: plupload on new-post
If a user is not logged in and they land on the new-post page plupload
is called but with a dev status null, breaking the js on the page.
Removed the restricted page message and redirected to /404/ page for
now.

Future: create login style page for redirects in the plugin